+++ lib/ooxml_parser/docx_parser/docx_data/document_structure.rb
@@ -122,7 +122,7 @@ module OoxmlParser
         end
         doc_defaults.xpath('w:rPrDefault').each do |r_pr_defaults|
           r_pr_defaults.xpath('w:rPr').each do |r_pr|
-            DocumentStructure.default_run_style = RunPropertiesDocument.parse(r_pr, DocxParagraphRun.new)
+            DocumentStructure.default_run_style = DocxParagraphRun.new.parse_properties(r_pr)
           end
         end
       end
@@ -181,11 +181,11 @@ module OoxmlParser
             DocxParagraph.parse_paragraph_style(paragraph_pr_tag, DocumentStructure.default_paragraph_style, DocumentStructure.default_run_style)
           end
           style.xpath('w:rPr').each do |character_pr_tag|
-            RunPropertiesDocument.parse(character_pr_tag, DocumentStructure.default_run_style, DocumentStructure.default_run_style)
+            DocumentStructure.default_run_style.parse_properties(character_pr_tag, DocumentStructure.default_run_style)
           end
         elsif (style.attribute('default').value == '1' || style.attribute('default').value == 'on' || style.attribute('default').value == 'true') && style.attribute('type').value == 'character'
           style.xpath('w:rPr').each do |character_pr_tag|
-            RunPropertiesDocument.parse(character_pr_tag, DocumentStructure.default_run_style, DocumentStructure.default_run_style)
+            DocumentStructure.default_run_style.parse_properties(character_pr_tag, DocumentStructure.default_run_style)
           end
         end
       end
@@ -199,7 +199,7 @@ module OoxmlParser
           DocxParagraph.parse_paragraph_style(table_paragraph_pr_tag, DocumentStructure.default_table_paragraph_style, DocumentStructure.default_table_run_style)
         end
         style.xpath('w:rPr').each do |table_character_pr_tag|
-          RunPropertiesDocument.parse(table_character_pr_tag, DocumentStructure.default_table_run_style, DocumentStructure.default_run_style)
+          DocumentStructure.default_table_run_style.parse_properties(table_character_pr_tag, DocumentStructure.default_run_style)
         end
       end
     end